// src/types/product.ts
export type Product = {
    index: number;
    tag: string;
    ram: string;
    memory: string;
    validity: string;
    price: number;
};


// export type Order = {
//     id: string|any;
//     date: Date;
//     value: number;
//     productId: number;
//     serverUname: string;
//     serverPassword: string;
// }